refactor: checkbox for 'I still work here' is now within the DateSelector component, with the states managed in that component
this helps reduce the number of props passed into the DateSelector functional component
