+++ src/components/DateSelector.jsx
@@ -1,15 +1,17 @@
 import { useState } from "react";
+import { Checkbox } from "@mantine/core";
 import { MonthPickerInput } from "@mantine/dates";
 
-export default function DateSelector({
-  isPresent = false,
-  isProfessionalExperience = false,
-}) {
+export default function DateSelector({ isProfessionalExperience = false }) {
   const [startDate, setStartDate] = useState(null);
   const [endDate, setEndDate] = useState(null);
+  const [positionIsPresent, setPositionIsPresent] = useState(false);
+
+  const handleCheckForPresentPosition = (event) =>
+    setPositionIsPresent(event.currentTarget.checked);
 
   return (
-    <>
+    <ul className="grid">
       <li>
         <MonthPickerInput
           label="Start Date"
@@ -25,16 +27,26 @@ export default function DateSelector({
         <MonthPickerInput
           label="End Date"
           key="endDate"
-          placeholder={isPresent ? "Present" : "Select month and year"}
+          placeholder={positionIsPresent ? "Present" : "Select month and year"}
           clearable
-          disabled={isPresent}
+          disabled={positionIsPresent}
           className="dateInput"
           minDate={startDate}
           maxDate={isProfessionalExperience && new Date()}
-          value={isPresent ? null : endDate}
+          value={positionIsPresent ? null : endDate}
           onChange={setEndDate}
         />
       </li>
-    </>
+      {isProfessionalExperience && (
+        <Checkbox
+          label="I still work here"
+          color="#00A9AD"
+          size="xs"
+          checked={positionIsPresent}
+          onChange={handleCheckForPresentPosition}
+          className="presentPositionCheckbox"
+        />
+      )}
+    </ul>
   );
 }
